move all modules into castor_api_client root module
