+++ castoredc_api_client/importer/helpers.py
@@ -7,7 +7,7 @@ from castoredc_api_client import CastorException
 
 
 if TYPE_CHECKING:
-    from castoredc_api_study import CastorStudy
+    from castoredc_api_client.study import CastorStudy
 
 
 def read_excel(path: str) -> pd.DataFrame:
