+++ castoredc_api_client/importer/import_data.py
@@ -6,10 +6,10 @@ from typing import TYPE_CHECKING, Dict, Optional
 from tqdm import tqdm
 
 from castoredc_api_client import CastorException
-from castoredc_api_import.helpers import create_upload, update_feedback
+from castoredc_api_client.importer.helpers import create_upload, update_feedback
 
 if TYPE_CHECKING:
-    from castoredc_api_study import CastorStudy
+    from castoredc_api_client.study import CastorStudy
 
 
 def import_data(
