+++ castoredc_api_client/study/castor_objects/castor_form.py
@@ -1,8 +1,8 @@
 import itertools
 from typing import List, Optional, Any, Union
 
-from castoredc_api_study.castor_objects.castor_field import CastorField
-from castoredc_api_study.castor_objects.castor_step import CastorStep
+from castoredc_api_client.study.castor_objects.castor_field import CastorField
+from castoredc_api_client.study.castor_objects.castor_step import CastorStep
 
 
 class CastorForm:
