+++ castoredc_api_client/study/castor_study.py
@@ -13,12 +13,12 @@ from tqdm import tqdm
 
 from castoredc_api_client import CastorException
 from castoredc_api_client.castoredc_api_client import CastorClient
-from castoredc_api_study.castor_objects.castor_data_point import CastorDataPoint
-from castoredc_api_study.castor_objects.castor_field import CastorField
-from castoredc_api_study.castor_objects.castor_form_instance import CastorFormInstance
-from castoredc_api_study.castor_objects.castor_record import CastorRecord
-from castoredc_api_study.castor_objects.castor_step import CastorStep
-from castoredc_api_study.castor_objects.castor_form import CastorForm
+from castoredc_api_client.study.castor_objects.castor_data_point import CastorDataPoint
+from castoredc_api_client.study.castor_objects.castor_field import CastorField
+from castoredc_api_client.study.castor_objects.castor_form_instance import CastorFormInstance
+from castoredc_api_client.study.castor_objects.castor_record import CastorRecord
+from castoredc_api_client.study.castor_objects.castor_step import CastorStep
+from castoredc_api_client.study.castor_objects.castor_form import CastorForm
 
 
 class CastorStudy:
