+++ tests/test_castor_objects/test_castor_form_instance.py
@@ -7,8 +7,8 @@ https://orcid.org/0000-0003-3052-596X
 """
 import pytest
 
-from castoredc_api_study.castor_objects.castor_data_point import CastorDataPoint
-from castoredc_api_study.castor_objects.castor_form_instance import CastorFormInstance
+from castoredc_api_client.study.castor_objects.castor_data_point import CastorDataPoint
+from castoredc_api_client.study.castor_objects.castor_form_instance import CastorFormInstance
 from castoredc_api_client import CastorException
