+++ tests/test_castor_objects/test_castor_record.py
@@ -5,8 +5,8 @@ Testing class for the CastorRecord class.
 @author: R.C.A. van Linschoten
 https://orcid.org/0000-0003-3052-596X
 """
-from castoredc_api_study.castor_objects.castor_form_instance import CastorFormInstance
-from castoredc_api_study.castor_objects.castor_record import CastorRecord
+from castoredc_api_client.study.castor_objects.castor_form_instance import CastorFormInstance
+from castoredc_api_client.study.castor_objects.castor_record import CastorRecord
 
 
 class TestCastorRecord:
