+++ tests/test_import/test_import_report.py
@@ -1,7 +1,7 @@
 import pytest
 
 from castoredc_api_client import CastorException
-from castoredc_api_import.import_data import import_data
+from castoredc_api_client.importer.import_data import import_data
 
 
 class TestImportReport:
