+++ tests/test_integration/fixtures_integration.py
@@ -1,7 +1,7 @@
 import pytest
 import auth.auth_data as auth_data
 
-from castoredc_api_study.castor_study import CastorStudy
+from castoredc_api_client.study.castor_study import CastorStudy
 
 
 @pytest.fixture(scope="class")
